+++ sensors/ADS131E.h
@@ -21,8 +21,8 @@ public:
     ~ADS131E() override;
     int32_t Initialize() override final;
     int32_t Uninitialize() override final;
-    int32_t Write(const void *data, uint32_t num) override final;
-    int32_t Read(void *data, uint32_t num) override final;
+    int32_t Write(const void *data, size_t num) override final;
+    int32_t Read(void *data, size_t num) override final;
     int32_t Control(uint32_t control, uint32_t arg) override final;
     virtual const char* Name() { return "ads131e"; }
